### 2021-02-04 #3006 @JulienPeloton
Hi @kyleniemeyer, @mbobra

I confirm this is the correct paper location. The paper lives in
- repo: https://github.com/JulienPeloton/s4cmb
- branch `issues/56/joss`
- folder: `paper/`

Note that it compiled correctly from the Whedon paper preview service ([link](https://whedon.theoj.org/)). Shall I move the files under the root of the repo instead of `paper/` to ease the review process?
