### 2021-02-08 #3022 @whedon
[REVIEW]: Simulating instrumental systematics of Cosmic Microwave Background experiments with s4cmb
**Submitting author:** @JulienPeloton (<a href="http://orcid.org/0000-0002-8560-4449">Julien Peloton</a>)
**Repository:** <a href="https://github.com/JulienPeloton/s4cmb" target ="_blank">https://github.com/JulienPeloton/s4cmb</a>
**Version:** v1.0
