### 2021-03-12 8042544
Update to describe how s4cmb compares to other commonly-used packages in this research area
