+++ paper/s4cmb_paper.bib
@@ -308,3 +308,17 @@ archivePrefix = {arXiv},
     year = "2018"
 }
 
+@software{theodore_kisner_2020_4270476,
+  author       = {Theodore Kisner and
+                  Reijo Keskitalo and
+                  Andrea Zonca and
+                  Giuseppe Puglisi and
+                  Julian Borrill},
+  title        = {hpc4cmb/toast: Small fixes discovered post 2.3.11},
+  month        = nov,
+  year         = 2020,
+  publisher    = {Zenodo},
+  version      = {2.3.12},
+  doi          = {10.5281/zenodo.4270476},
+  url          = {https://doi.org/10.5281/zenodo.4270476}
+}
