### 2021-03-15 31d2080
Update README based on JOSS reviewer suggestion. Thanks @Christovis!
