### 2021-04-05 #3022 @mbobra
@JulienPeloton Are you willing to version this software using semantic versioning? It is not required, but it is a nice standard. Please let me know.
