### 2021-04-05 #3022 @JulienPeloton
Hi @mbobra -- s4cmb is already versioned using semantic versioning, see e.g.:
- https://github.com/JulienPeloton/s4cmb/releases
- https://pypi.org/project/s4cmb/#history

Only the Zenodo archive did not have a number, but it corresponds to version `1.0` of the library.
