### 2021-04-05 #3022 @mbobra
@JulienPeloton I see. Could you then change Zenodo archive for the JOSS paper to v1.0?
