### 2021-04-05 #3022 @danielskatz
@JulienPeloton - can you change the metadata in the repository so that the title is "Simulating instrumental systematics of Cosmic Microwave Background experiments with s4cmb", matching the paper (this doesn't require a new deposit or new version, just the metadata changes)
