### 2021-04-05 #3022 @danielskatz
A bunch of small changes to the paper can be found in https://github.com/JulienPeloton/s4cmb/pull/58
