### 2021-04-06 #3022 @JulienPeloton
@danielskatz -- thanks for the proofreading, the changes have been merged.

I have also edited the Zenodo archive with the same title as the paper (https://zenodo.org/record/4659617).
